KSENROLL-2245, KSENROLL-2292.  Tested 2245.  Seems like it works in unit-test land.  Had to modify getRegistrationGroup to do a null test and throw exception.  Prelim check-in for 2292.  Still need to test and wire.
git-svn-id: https://svn.kuali.org/repos/student/branches/ks-1.3@34889 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-enroll/ks-enroll-impl/src/main/java/org/kuali/student/enrollment/class2/courseoffering/service/impl/CourseOfferingServiceImpl.java
@@ -1393,6 +1393,9 @@ public class CourseOfferingServiceImpl implements CourseOfferingService {
     public RegistrationGroupInfo getRegistrationGroup(String registrationGroupId, ContextInfo context) throws DoesNotExistException, InvalidParameterException, MissingParameterException,
             OperationFailedException, PermissionDeniedException {
         LuiInfo lui = luiService.getLui(registrationGroupId, context);
+        if (lui == null) {
+            throw new DoesNotExistException("registrationGroupId does not exist: " + registrationGroupId);
+        }
         RegistrationGroupInfo rgInfo = registrationGroupTransformer.lui2Rg(lui, context);
         rgInfo.setCourseOfferingId(this.getFormatOffering(rgInfo.getFormatOfferingId(), context).getCourseOfferingId());
 
@@ -1543,6 +1546,14 @@ public class CourseOfferingServiceImpl implements CourseOfferingService {
     public StatusInfo deleteRegistrationGroup(String registrationGroupId, ContextInfo context) throws DoesNotExistException, InvalidParameterException, MissingParameterException,
             OperationFailedException, PermissionDeniedException {
         try {
+            LuiInfo fetch = luiService.getLui(registrationGroupId, context);
+            if (fetch == null) {
+                throw new DoesNotExistException("Registration Group, " + registrationGroupId + ", does not exist");
+            }
+            // Make sure we have correct type before deleting
+            if (!LuiServiceConstants.REGISTRATION_GROUP_TYPE_KEY.equals(fetch.getTypeKey())) {
+                throw new InvalidParameterException("ID, " + registrationGroupId + ", does not have a registration group type");
+            }
             return luiService.deleteLui(registrationGroupId, context);
         } catch (DependentObjectsExistException e) {
             throw new OperationFailedException("Could not delete LUI '" + registrationGroupId + "'", e);
